% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{add_grid_table}
\alias{add_grid_table}
\title{Add a grid table to a GTFS feed}
\usage{
add_grid_table(gtfs, grid)
}
\arguments{
\item{gtfs}{An object of class \link[gtfsio:import_gtfs]{gtfs}.}

\item{grid}{A \code{\link[data.table:data.table]{data.table::data.table()}} object created with
created with \code{\link[=create_grid_table]{create_grid_table()}}.}
}
\value{
An object of class \link[gtfsio:import_gtfs]{gtfs}. This is a
list of multiple \code{\link[data.table:data.table]{data.table::data.table()}} objects corresponding
to the different dataset files of a GTFS feed. In addition to those, the
returned object will contain a \code{data.table} representing the reference
grid of a GTFS-Multi feed. See details.
}
\description{
Add a grid table to a GTFS feed
}
\details{
The grid table of a GTFS-Multi feed always has the same structure
as the \href{https://developers.google.com/transit/gtfs/reference#stopstxt}{stops}
table in a regular GTFS feed. Instead of a true transit stop location, each
row represents a grid point (i.e. the centroid of a pixel in the reference
grid).
}
\seealso{
\code{\link[=create_grid_table]{create_grid_table()}}
}
