% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{add_transfer_tables}
\alias{add_transfer_tables}
\title{Add a transfer table to a GTFS feed}
\usage{
add_transfer_tables(gtfs, ...)
}
\arguments{
\item{gtfs}{An object of class \link[gtfsio:import_gtfs]{gtfs}.}

\item{...}{One or more keyword arguments, with the key being the name of a
specific transport mode, and the value being a
\code{\link[data.table:data.table]{data.table::data.table()}} object containing the travel times
between different transit stop locations for that mode. Such a table can be
created with \code{\link[=create_transfer_table]{create_transfer_table()}}, using transit stop
locations as both origins and destinations.}
}
\value{
An object of class \link[gtfsio:import_gtfs]{gtfs}. This is a
list of multiple \code{\link[data.table:data.table]{data.table::data.table()}} objects corresponding
to the different dataset files of a GTFS feed. In addition to those, the
returned object will contain a \code{data.table} representing the transfer
trips of a GTFS-Multi feed. See details.
}
\description{
Add a transfer table to a GTFS feed
}
\details{
The transfer table of a GTFS-Multi feed always has the same
structure as the
\href{https://developers.google.com/transit/gtfs/reference#transferstxt}{transfer}
table in a regular GTFS feed. Each row contains
the minimum travel time needed to travel between two different transit stop
locations, using a certain transport mode. For each
transport mode, corresponding travel times are stored in a separate column.
}
\note{
A transfer table already present in the input gtfs object will be
overwritten.
}
\seealso{
\code{\link[=create_transfer_table]{create_transfer_table()}}
}
