% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{assert_multinet}
\alias{assert_multinet}
\title{Assert a GTFS feed is a valid GTFS-Multi feed}
\usage{
assert_multinet(x)
}
\arguments{
\item{x}{An object of class \code{multinet} or \code{gtfs}.}
}
\value{
Returns nothing if x is a valid GTFS-Multi feed, throws and error
otherwise.
}
\description{
Assert a GTFS feed is a valid GTFS-Multi feed
}
\details{
Currently it is only checked if all required datasets are present.
}
