% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{build_streetnet}
\alias{build_streetnet}
\title{Build a routable graph from a set of linestrings}
\usage{
build_streetnet(x, protect = NULL)
}
\arguments{
\item{x}{An object of class \code{\link[sf:sf]{sf}} with \code{LINESTRING} or
\code{MULTILINESTRING} geometries.}

\item{protect}{Set of points that protect their nearest linestring vertex
from being smoothed (see Details). That is, the closest linestring vertices
to these points are guaranteed to be a node in the resulting graph, even if
they are neither a terminal node nor a junction node. May be given either as
an object of class \code{\link[sf:sf]{sf}} or \code{\link[sf:st_as_sfc]{sfc}} with \code{POINT}
geometries, or as a two-column numeric matrix containing respectively the
longitude and latitude coordinates of the points.}
}
\value{
An object of class \code{\link[sfnetworks:sfnetwork]{sfnetworks}}.
}
\description{
Build a routable graph from a set of linestrings
}
\details{
Graph building is implemented as a step-wise process. The graph
will be directed, so first all linestring geometries of \code{x} that are \emph{not}
marked as being a oneway street (through the \emph{oneway} attribute column if
present) are duplicated and reversed. Then, all the linestring geometries
are split such that their segments become separate linestring geometries
on their own. These serve as the edges in the initial graph, and their
endpoints become the nodes. Endpoints that are shared between multiple edges
become a single node, such that these edges are connected. Only the largest
connected component of the initial graph is preserved. Finally, the initial
graph is smoothed by removing all nodes that are neither a terminal node
(i.e. a node at the end of an edge without a connection to any other edge)
nor a junction node (i.e. a node that connects more than two edges) nor a
nearest node to any of the points passed to the \code{protect} parameter.
}
\note{
Coordinates should be expressed in coordinate reference system
\href{https://epsg.io/4326}{EPSG:4326}. The function assumes this to be true, and
does not check for it.
}
