% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{create_extent}
\alias{create_extent}
\title{Create a rectangular spatial extent for the analysis}
\usage{
create_extent(bounds, input_crs = 4326, output_crs = input_crs)
}
\arguments{
\item{bounds}{Coordinates marking the boundaries of the extent, as an
ordered vector in the form \code{c(xmin, ymin, xmax, ymax)}.}

\item{input_crs}{Coordinate reference system in which the coordinate values
of \code{bounds} are expressed. Can be anything understood by \code{\link[sf:st_crs]{sf::st_crs()}}. By
default equal to 4326, i.e. longitude and latitude coordinates with WGS84 as
geodetic datum.}

\item{output_crs}{Coordinate reference system to create the extent in. Can
be anything understood by \code{\link[sf:st_crs]{sf::st_crs()}}. By default equal to \code{input_crs}.}
}
\value{
The rectangular extent as a single \code{POLYGON} geometry in an object
of class \link[sf:sfc]{sfc}.
}
\description{
Create a rectangular spatial extent for the analysis
}
