% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{create_grid_table}
\alias{create_grid_table}
\title{Create a grid table for a reference grid}
\usage{
create_grid_table(grid)
}
\arguments{
\item{grid}{The reference grid as object of class
\code{\link[stars:st_as_stars]{stars}}.}
}
\value{
A \code{\link[data.table:data.table]{data.table}} object. See details.
}
\description{
Create a grid table for a reference grid
}
\details{
The grid table of a GTFS-Multi feed always has the same structure
as the \href{https://developers.google.com/transit/gtfs/reference#stopstxt}{stops}
table in a regular GTFS feed. Instead of a true transit stop location, each
row represents a grid point (i.e. the centroid of a pixel in the reference
grid).
}
\seealso{
\code{\link[=create_reference_grid]{create_reference_grid()}}
}
