% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_reference_grid}
\alias{create_reference_grid}
\title{Create a reference grid over the spatial extent of the analysis}
\usage{
create_reference_grid(extent, cellsize, output_crs = sf::st_crs(extent))
}
\arguments{
\item{extent}{The spatial extent of the analysis as a single \code{POLYGON}
geometry in either object of class \code{\link[sf:sf]{sf}} or
\code{\link[sf:sfc]{sfc}}.}

\item{cellsize}{Cellsize in units of \code{output_crs} (usually meters). Cells
are always squares, i.e. the width and height of the cells are both equal
to \code{cellsize}.}

\item{output_crs}{CRS to create the grid in. Can be anything understood by
\code{\link[sf:st_crs]{sf::st_crs()}}. By default equal to the CRS of \code{extent}.}
}
\value{
The reference grid as an object of class
\code{\link[stars:st_as_stars]{stars}}.
}
\description{
Create a reference grid over the spatial extent of the analysis
}
\seealso{
\code{\link[=create_extent]{create_extent()}}, \code{\link[=create_grid_table]{create_grid_table()}}
}
