% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{create_transfer_table}
\alias{create_transfer_table}
\title{Create a transfer table for a set of origins and destinations}
\usage{
create_transfer_table(
  origins,
  destinations,
  streetnet,
  d_limit = Inf,
  min_time = 120,
  ...
)
}
\arguments{
\item{origins}{A \code{\link[data.table:data.table]{data.table::data.table()}} object containing the
origin locations of the transfer trips. See details.}

\item{destinations}{A \code{\link[data.table:data.table]{data.table::data.table()}} object containing
the destination locations of the transfer trips. See details.}

\item{streetnet}{The street network as an object of class
\code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork()}} or \code{\link[dodgr:dodgr_streetnet]{dodgr::dodgr_streetnet()}}.}

\item{d_limit}{Upper distance limit (as-the-crow-flies) in meters to consider
a possible transfer between an origin and a destination.}

\item{min_time}{Minimum travel time for any transfer trip in seconds,
independent of the calculated travel time by the routing algorithm.}

\item{...}{Additional parameters passed on to \code{\link[=streetnet_traveltimes]{streetnet_traveltimes()}}.
When parameter \code{streetnet} is of class \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork()}}, these
include the \code{time_column} and \code{weight_column} parameters. See
\code{\link[=streetnet_traveltimes]{streetnet_traveltimes()}} for details.}
}
\description{
Create a transfer table for a set of origins and destinations
}
\details{
Both origins and destinations should be given in the format of a
\href{https://developers.google.com/transit/gtfs/reference#stopstxt}{stops} table
of a GTFS feed.

The format of a
\href{https://developers.google.com/transit/gtfs/reference#transferstxt}{transfer}
table can be used to model both true transit transfer trips as well as
access, direct and egress trips:
\itemize{
\item When modelling access trips, the origins will be grid points and the
destinations transit stop locations.
\item When modelling egress trips the origins will be transit stop locations and
the destinations grid points.
\item When modelling direct trips both origins and destinations will be grid
points.
\item When modelling true transit transfer trips, both origins and destinations
will be transit stop locations.
}
}
\seealso{
\code{\link[=add_access_tables]{add_access_tables()}}, \code{\link[=add_direct_tables]{add_direct_tables()}},
\code{\link[=add_egress_tables]{add_egress_tables()}}, \code{\link[=add_transfer_tables]{add_transfer_tables()}}
}
