% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_traveltime_grid}
\alias{create_traveltime_grid}
\title{Write travel times as attributes to a grid}
\usage{
create_traveltime_grid(x, grid, dimname = "time")
}
\arguments{
\item{x}{The object returned by \code{\link[=multinet_traveltimes]{multinet_traveltimes()}}, i.e. the grid
table of a GTFS-Multi feed as object of class
\code{\link[data.table:data.table]{data.table}} with one or multiple columns
containing calculated travel time values. These columns should have names
that start with "travel_time".}

\item{grid}{Reference grid of \code{x} as object of class
\code{\link[stars:st_as_stars]{stars}}.}

\item{dimname}{If \code{x} contains multiple columns containing travel time
values, they will be stored along an additional dimension. How should this
dimension be called? Defaults to "time", assuming that the different columns
refer to different departure times.}
}
\value{
The travel time grid as an object of class
\code{\link[stars:st_as_stars]{stars}}. This grid has the same spatial dimensions
as \code{x} but contains only travel time values for each pixel.
}
\description{
Write travel times as attributes to a grid
}
\seealso{
\code{\link[=multinet_traveltimes]{multinet_traveltimes()}}, \code{\link[=create_reference_grid]{create_reference_grid()}}
}
