% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\docType{data}
\name{DEFAULT_HIGHWAY_TYPES}
\alias{DEFAULT_HIGHWAY_TYPES}
\title{Default highway types to be included in OpenStreetMap street import}
\format{
An object of class \code{character} of length 21.
}
\usage{
DEFAULT_HIGHWAY_TYPES
}
\description{
Default highway types to be included in OpenStreetMap street import
}
\seealso{
\code{\link[=import_streetnet]{import_streetnet()}}
}
\keyword{datasets}
