% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\docType{data}
\name{DEFAULT_TAGS}
\alias{DEFAULT_TAGS}
\title{Default tags to be included as columns in OpenStreetMap street import}
\format{
An object of class \code{character} of length 6.
}
\usage{
DEFAULT_TAGS
}
\description{
Default tags to be included as columns in OpenStreetMap street import
}
\seealso{
\code{\link[=import_streetnet]{import_streetnet()}}
}
\keyword{datasets}
