% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{export_grid}
\alias{export_grid}
\title{Write a grid to disk}
\usage{
export_grid(x, path, layer = 1, quiet = FALSE)
}
\arguments{
\item{x}{The grid as object of class \code{\link[stars:st_as_stars]{stars}}.}

\item{path}{Path to a \code{.geotiff} or \code{tiff} file where the grid
should be written to.}

\item{layer}{Name or index of the attribute to be written. Defaults to 1,
meaning the first attribute is written.}

\item{quiet}{Boolean. If \code{FALSE}, display progress information on screen.
Defaults to \code{FALSE}.}
}
\description{
Write a grid to disk
}
\seealso{
\code{\link[=create_reference_grid]{create_reference_grid()}}, \code{\link[=create_traveltime_grid]{create_traveltime_grid()}}
}
