% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{export_multinet}
\alias{export_multinet}
\title{Write a GTFS-Multi feed to disk}
\usage{
export_multinet(x, path, ...)
}
\arguments{
\item{x}{An object of class \code{multinet}, which is a list of multiple
\code{\link[data.table:data.table]{data.table}} objects corresponding to the
different dataset files of a GTFS-Multi feed.}

\item{path}{Path to a \code{.zip} file where the GTFS-Multi feed should be
written to.}

\item{...}{Additional arguments passed on to \code{\link[gtfsio:export_gtfs]{gtfsio::export_gtfs()}}.}
}
\description{
Write a GTFS-Multi feed to disk
}
