% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_neighbours}
\alias{get_neighbours}
\title{Select neighbouring locations of a given origin}
\usage{
get_neighbours(origin, candidates, d_limit = Inf)
}
\arguments{
\item{origin}{A numeric vector of length two containing respectively
the longitude and latitude coordinate of the origin location.}

\item{candidates}{A \code{\link[=data.frame]{data.frame()}} or similar representing the set of
possible neighbour locations to choose from. Should have only three columns,
with the first column containing unique indices for each location, and the
second and third column containing respectively the longitude and latitude
coordinates of each location.}

\item{d_limit}{Upper distance limit (as-the-crow-flies) in meters to
consider a candidate location as a neighbour of the origin. Defaults to
\code{Inf}, meaning that each candidate location will be a neighbour.}
}
\value{
A subset of \code{candidates}.
}
\description{
Select neighbouring locations of a given origin
}
\note{
Coordinates should be expressed in coordinate reference system
\href{https://epsg.io/4326}{EPSG:4326}. The function assumes this to be true, and
does not check for it.
}
