% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{import_grid}
\alias{import_grid}
\title{Import a grid from a GeoTIFF file}
\usage{
import_grid(path, quiet = FALSE)
}
\arguments{
\item{path}{Path to a \code{.geotiff} or \code{.tiff} file containing the grid.}

\item{quiet}{Boolean. If \code{FALSE}, display progress information on screen.
Defaults to \code{FALSE}.}
}
\value{
The grid as an object of class \code{\link[stars:st_as_stars]{stars}}.
}
\description{
Import a grid from a GeoTIFF file
}
\seealso{
\code{\link[=create_reference_grid]{create_reference_grid()}}, \code{\link[=create_traveltime_grid]{create_traveltime_grid()}}
}
