% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{import_multinet}
\alias{import_multinet}
\title{Import a GTFS-Multi feed from a zip archive}
\usage{
import_multinet(path)
}
\arguments{
\item{path}{Path to a \code{.zip} file containing all required GTFS-Multi
datasets as \code{.txt} files.}
}
\value{
An object of class \code{multinet}, which is a list of multiple
\code{\link[data.table:data.table]{data.table::data.table()}} objects corresponding to the different dataset
files of a GTFS-Multi feed.
}
\description{
Import a GTFS-Multi feed from a zip archive
}
