% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{import_streetnet}
\alias{import_streetnet}
\title{Import streets from an OpenStreetMap extract}
\usage{
import_streetnet(
  path,
  extent = NULL,
  clip = FALSE,
  highway_types = DEFAULT_HIGHWAY_TYPES,
  tags = DEFAULT_TAGS,
  quiet = FALSE
)
}
\arguments{
\item{path}{Path to a \code{.osm.pbf} file containing the extracted OpenStreetMap
data for at least the extent of the area of interest.}

\item{extent}{The area of interest as object of class \code{\link[sf:sf]{sf::sf()}}. Only
streets that intersect this area will be imported. If \code{NULL}, streets are
imported for the complete OpenStreetMap extract.}

\item{clip}{Boolean. If \code{TRUE}, linestrings will be clipped by the given
extent. If \code{FALSE}, linestrings that intersect the given extent will be
imported as a whole. Defaults to \code{FALSE}. Ignored when \code{extent = NULL}.}

\item{highway_types}{A vector of possible values for OpenStreetMaps
\href{https://wiki.openstreetmap.org/wiki/Key:highway}{highway tag}. Only streets
that have one of these values for the highway tag will be imported. Defaults
to \link{DEFAULT_HIGHWAY_TYPES}.}

\item{tags}{A vector of OpenStreetMap
\href{https://wiki.openstreetmap.org/wiki/Tags}{tags} that should be included as
attribute columns in the imported data. Defaults to \link{DEFAULT_TAGS}.}

\item{quiet}{Boolean. If \code{FALSE}, display progress information on screen.
Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link[sf:sf]{sf::sf()}} with \code{LINESTRING} or
\code{MULTILINESTRING} geometries.
}
\description{
Import streets from an OpenStreetMap extract
}
