% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{multinet_to_gtfs}
\alias{multinet_to_gtfs}
\title{Convert a GTFS-Multi feed to a regular GTFS feed}
\usage{
multinet_to_gtfs(
  x,
  access_modes = NULL,
  direct_modes = NULL,
  egress_modes = NULL,
  transfer_modes = NULL,
  access_model = "transfers",
  direct_model = "transfers",
  egress_model = "transfers",
  access_frequency = 60,
  direct_frequency = 60,
  egress_frequency = 60,
  time_as_seconds = FALSE
)
}
\arguments{
\item{x}{An object of class \code{multinet}, which is a list of multiple
\code{\link[data.table:data.table]{data.table::data.table()}} objects corresponding to the different dataset
files of a GTFS-Multi feed.}

\item{access_modes}{Vector of transport modes to consider for access trips,
i.e. trips from grid points towards transit stop locations.
Each of these transport modes should have their access travel time for each
access trip specified in the access table of \code{x}, in a column named
\emph{transfer_time_{<}mode{>}}. If \code{NULL}, all modes that have such a column
in the access table are considered. Defaults to \code{NULL}.}

\item{direct_modes}{Vector of transport modes to consider for direct trips,
i.e. trips between grid points.
Each of these transport modes should have their direct travel time for each
direct trip specified in the direct table of \code{x}, in a column named
\emph{transfer_time_{<}mode{>}}. If \code{NULL}, all modes that have such a column
in the direct table are considered. Defaults to \code{NULL}.}

\item{egress_modes}{Vector of transport modes to consider for egress trips,
i.e. trips from transit stop locations towards grid points.
Each of these transport modes should have their egress travel time for each
egress trip specified in the egress table of \code{x}, in a column named
\emph{transfer_time_{<}mode{>}}. If \code{NULL}, all modes that have such a column
in the egress table are considered. Defaults to \code{NULL}.}

\item{transfer_modes}{Vector of transport modes to consider for true transit
transfer trips, i.e. trips between transit stop locations.
Each of these transport modes should have their transfer travel time for
each transfer trip specified in the transfer table of \code{x}, in a column named
\emph{transfer_time_{<}mode{>}}. If \code{NULL}, all modes that have such a column
in the transfer table are considered. Defaults to \code{NULL}.}

\item{access_model}{How to model access trips, i.e. trips from grid points
towards transit stop locations. Can either be "transfers" to model them as
if they where transfers between transit stops or "trips" to model them as
if they where regular transit trips departing at a certain frequency. This
frequency is given in the \code{access_frequency} parameter. May also be "ignore"
to not include them at all in the GTFS feed.}

\item{direct_model}{How to model direct trips, i.e. trips between grid
points. Can either be "transfers" to model them as if they where
transfers between transit stops or "trips" to model them as if they
where regular transit trips departing at a certain frequency. This
frequency is given in the \code{direct_frequency} parameter. May also be "ignore"
to not include them at all in the GTFS feed.}

\item{egress_model}{How to model egress trips, i.e. trips from transit stop
locations towards grid points. Can either be "transfers" to model them as
if they where transfers between transit stops or "trips" to model them as
if they where regular transit trips departing at a certain frequency. This
frequency is given in the \code{egress_frequency} parameter. May also be "ignore"
to not include them at all in the GTFS feed.}

\item{access_frequency}{If access trips should be modelled as
frequency-based transit trips (i.e. if \code{access_model = "trips"}), at what
frequency should these trips operate? Value should be given in seconds.
Smaller values will provide more accurate results, since access trips can
in reality be started at anytime, but also increase processing time.
Defaults to 60, i.e. every minute.}

\item{direct_frequency}{If direct trips should be modelled as
frequency-based transit trips (i.e. if \code{direct_model = "trips"}), at what
frequency should these trips operate? Value should be given in seconds.
Smaller values will provide more accurate results, since direct trips can
in reality be started at anytime, but also increase processing time.
Defaults to 60, i.e. every minute.}

\item{egress_frequency}{If egress trips should be modelled as
frequency-based transit trips (i.e. if \code{egress_model = "trips"}), at what
frequency should these trips operate? Value should be given in seconds.
Smaller values will provide more accurate results, since egress trips can
in reality be started at anytime, but also increase processing time.
Defaults to 60, i.e. every minute.}

\item{time_as_seconds}{Should arrival and departure times be stored as
seconds since midnight? If \code{FALSE}, they are stored as regular
time strings in format "h:m:s". Defaults to \code{FALSE}.}
}
\value{
An object of class \link[gtfsio:import_gtfs]{gtfs}, which is a
list of multiple \code{\link[data.table:data.table]{data.table::data.table()}} objects corresponding
to the different dataset files of a GTFS feed.
}
\description{
Convert a GTFS-Multi feed to a regular GTFS feed
}
