% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{reduce_to_percentiles}
\alias{reduce_to_percentiles}
\alias{reduce_to_percentiles.data.table}
\alias{reduce_to_percentiles.stars}
\title{Reduce a multiple travel times to a set of percentile values}
\usage{
reduce_to_percentiles(x, percentiles = c(5, 25, 50, 75, 95))

\method{reduce_to_percentiles}{data.table}(x, percentiles = c(5, 25, 50, 75, 95))

\method{reduce_to_percentiles}{stars}(x, percentiles = c(5, 25, 50, 75, 95))
}
\arguments{
\item{x}{Object to reduce. May be a travel time grid as object of class
\code{\link[stars:st_as_stars]{stars}} created by \code{\link[=create_traveltime_grid]{create_traveltime_grid()}} and
having multiple travel time values per spatial location stored along a third
dimension. May also be an object of class
\code{\link[data.table:data.table]{data.table}} created by \code{\link[=multinet_traveltimes]{multinet_traveltimes()}}
and having multiple travel time columns with names starting with
"travel_time".}

\item{percentiles}{Percentile values that should be calculated.}
}
\value{
In case \code{x} is a travel time grid: the updated travel time grid as
object of class \code{\link[stars:st_as_stars]{stars}}, with the third now having
one coordinate per calculated percentile value. The original third
dimension is removed. In case \code{x} is a data table: the updated table having
one travel time column per calculated percentile, all named "travel_time_p"
followed by the corresponding percentile value. The original travel time
columns are removed.
}
\description{
Reduce a multiple travel times to a set of percentile values
}
\seealso{
\code{\link[=create_traveltime_grid]{create_traveltime_grid()}}, \code{\link[=multinet_traveltimes]{multinet_traveltimes()}}
}
