% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetnet.R
\name{streetnet_traveltimes}
\alias{streetnet_traveltimes}
\alias{streetnet_traveltimes.sfnetwork}
\alias{streetnet_traveltimes.dodgr_streetnet}
\alias{streetnet_traveltimes.dodgr_streetnet_sc}
\title{Calculate travel times on a street network}
\usage{
streetnet_traveltimes(x, od_pairs, ...)

\method{streetnet_traveltimes}{sfnetwork}(
  x,
  od_pairs,
  time_column = "time",
  weight_column = "weight",
  ...
)

\method{streetnet_traveltimes}{dodgr_streetnet}(x, od_pairs, ...)

\method{streetnet_traveltimes}{dodgr_streetnet_sc}(x, od_pairs, ...)
}
\arguments{
\item{x}{The street network as object of class \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork()}},
\code{\link[dodgr:dodgr_streetnet]{dodgr::dodgr_streetnet()}} or
\link[dodgr:dodgr_streetnet]{dodgr::dodgr_streetnet_sc()}.}

\item{od_pairs}{A \code{\link[data.table:data.table]{data.table::data.table()}} object in which each row
contains the origin and destination location of a route for which the travel
time needs to be calculated. The table should contain the columns
\emph{from_lon}, \emph{from_lat}, \emph{to_lon} and \emph{to_lat} containing respectively the
longitude and latitude coordinates of respectively the origin and
destination.}

\item{...}{Ignored.}

\item{time_column}{Name of the column in the edges table of the street
network that contains the travel time values for each edge.}

\item{weight_column}{Name of the column in the edges table of the street
network that contains the weight of each edge. These value are used to find
the shortest path (i.e. the path with minimum weight) between an origin and
a destination.}
}
\value{
An integer vector of travel times in seconds for each of the given
OD pairs.
}
\description{
Calculate travel times on a street network
}
\note{
Coordinates should be expressed in coordinate reference system
\href{https://epsg.io/4326}{EPSG:4326}. The function assumes this to be true, and
does not check for it.
}
